txList refactoring pull to refresh logic improvement list refresh logic improvements
